CPID/cnc-app/backend/pyproject.toml
[tool.poetry]
name = "cnc"
version = "0.1.0"
description = ""
authors = ["chenyili-jg"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
django = "3.2.12"
djangorestframework = "3.12.4"
blueapps = "4.1.1"
celery = "4.4.7"
urllib3 = "1.26.9"
eventlet = "0.31.0"
django-celery-beat = "2.2.0"
django-celery-results = "2.0.1"
django-cors-headers = "3.7.0"
mysqlclient = "2.0.3"
gunicorn = "20.1.0"

[tool.poetry.dev-dependencies]
black = "==22.3.0"
isort = "^5.10.1"
flake8 = "==6.0.0"
flake8-bugbear = "==23.1.20"
flake8-comprehensions = "==3.10.1"
flake8-pie = "==0.16.0"
filelock = "*"
mypy = "==1.0.0"
pylint = "==2.13.4"
pytest = "==6.2.5"
pytest-xdist = "*"
pytest-cov = "*"
pytest-django = "==4.5.2"
pytest-mock = "==3.2.0"
types-pyyaml = "^5"
types-redis = "^4.1.18"
types-dataclasses = "^0.6.4"
types-pytz = "^2021.3.6"
types-requests = "^2.27.15"
types-six = "^1.16.12"
types-bleach = "^4.1.5"
types-pymysql = "^1.0.14"
types-chardet = "^4.0.3"
types-python-dateutil = "^2.8.10"

[tool.black]
skip-string-normalization = 'true'
line-length = 119
exclude = '''
(
  /(
  .*/migrations
  | home_application
  | blueking
  | config
  )/
)
'''

[tool.isort]
profile = "black"
force_grid_wrap = 0
include_trailing_comma = true
line_length = 119
multi_line_output = 3
use_parentheses = true
skip_glob = [".mypy_cache", ".git", ".*/migrations", "blueking"]

[tool.flake8]
max-line-length=119
ignore = "F405,F403,W504,E741,E125,W503,F841,E203,E231,F541,C417,C901,F401,W391"
max-complexity=12
format = "pylint"
show_source = "true"
statistics = "true"
count = "true"
exclude = "*migrations*,*.pyc,.git,__pycache__,settings.py,blueking"

[tool.mypy]
ignore_missing_imports = true
show_error_codes = true
check_untyped_defs = true
namespace_packages = true
exclude = ["blueking", "config", "settings.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
